+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 4.7
 --------------
 - Clean code: remove else when not necessary in functions/*
+- SQL use NOT EXISTS(SELECT...) instead of NOT IN(SELECT...) in MassCreateAssignments.php
 
 Changes in 4.6
 --------------